Refactored lib/auth/services/auth_service.dart -> lib/auth/services/auth_services/auth_service.dart
